+++ packages/terra-site/CHANGELOG.md
@@ -7,11 +7,13 @@ Unreleased
 1.20.0 - (December 5, 2017)
 ------------------
 ### Added
+* Add terra-time-input mock theme styles and 12 hour input examples.
 * Added styles for terra-modal content displaying div.
 
 ### Changed
 * Minor version bump
 * Place site header background on a parent div, instead of collapsible.
+
 
 1.19.0 - (November 28, 2017)
 ------------------